fix naming and few env variable fixes
