+++ src/model/ProtocolConstants.ts
@@ -4,6 +4,7 @@ export interface ProtocolConstants {
   creditTokenAddress: string;
   profitManagerAddress: string;
   lendingTermOffboardingAddress: string;
+  lendingTermOnboardingAddress: string;
   uniswapV2RouterAddress: string;
   gatewayAddress: string;
   psmAddress: string;
