+++ src/utils/TelegramHelper.ts
@@ -1,9 +1,6 @@
 import axios from 'axios';
 import { sleep } from './Utils';
 
-const TG_BOT_ID: string | undefined = process.env.TG_BOT_ID;
-const TG_CHAT_ID: string | undefined = process.env.TG_CHAT_ID;
-
 let lastTGCall = Date.now();
 type TGBody = {
   chat_id: string;
@@ -11,6 +8,8 @@ type TGBody = {
   parse_mode?: string;
 };
 async function CallTelegram(msg: string, isMarkdown = false) {
+  const TG_BOT_ID: string | undefined = process.env.TG_BOT_ID;
+  const TG_CHAT_ID: string | undefined = process.env.TG_CHAT_ID;
   if (!TG_CHAT_ID || !TG_BOT_ID) {
     return;
   }
